Overhaul internals to support masked chars in raw
The aim of this change is to support code typing texts using the quote
function.

This commit changes the previous rendering workflow to one that only
runs changes in the necesary words when there is a render-worthy event.
Now the screen uses cells extensively which have been extended to
contain format metadata. The tracking is changed to a new set of strings
in the evaluatio where "typed" is compared to "rubric" but rendering and
cursor placement are based on "text" which can contain masked characters.

99% of functionality works fine. Styling and backspace in masked
characters in raw mode is still wonky.

Further details in PR